Added closest feature
I added the closest function which will determine which number based on a set of values (values) is closest to  a different given value (x). This function later needs to be added to the actual network or (read TODO).

Hope this can be a sign of our starting of the project.

Regards Mattis