Fen to bitboard decoder is working
After finding the 1 bug that has been challenging me and my metal health for the past hour (i initialised the counter variable field as 0 in the mathc loop) I can finally commit this part to the training branch. I did some inspecting and everything is working smoothly.
The bitboard has the composition according to what is stated in the comment for the fen_to_bitboard func.
However the side the black pieces start on is now on the top. This has to be changed according to how we need the output. This wont be a problem tho as the individual fields of the bitboard only need to be transposed (which is a numpy intern matrix operation) in order to have white pieces on the other side while keeping the correct composition of the pieces (kings and queens on the right square).
Enough said.
Regards Mattis