Started refactoring train_model.py
Started refactoring the train_model.py, so it is basically only 1 function which can be used for training the neural network. The type of dataset will determine if it is a neural network for white moves or for black moves. 

Due to some issues with my venv I have only been able to refactor the way of getting the batch size by now using the envisaged function __len__ of a dataset to get the whole size of the training data. 

Still a lot of work is due and if there may be any questions or concerns regarding the way I'm restructuring this file please contact me. 
Have a good evening, 
Regards, Mattis!