Added prepare_chess_data function which is meant to load the data from the respective CSV files.
Due to issues with the pace at which pandas seems to be loading each row of a CSV file the function will load a limited number of rows and not the whole dataset. So in order to be able to skip parts of the dataset which have already been covered the iter variable may be used as a measure to index the dataset into sectors with the size of the batch. This makes us able to later on try around with different batch sizes.

Regards Mattis

Signed-off-by: mattis <mattis.backhaus@maxe-del.de>