Added basic way of generating an output for the AI by creating a so-called test dataset which only contains 1 game state. This makes us be able to use the Dataloader for getting the training data. Therefore, this current method is quite simple.
What is still missing is the interpretation of the given output as right now its only numbers and not actual sans or letters. However, as I am still planning to rework the decode process for the given annotations I have not added that part so far. Using Unicode might introduce us to certain problems for instance the network always generating higher numbers for the last few digits of the annotation as these usually contain letters like x which in numeric value are quite high.

Tl;dr I am reworking the decoding process for the annotations based on their frequency.

Signed-off-by: mattis <mattis.backhaus@maxe-del.de>