Fixed a bug in chess_csv.py
The create outputs script was using the wrong column of the csv files for getting the moves. This bug likely came to life when we changed the dimension of the csv files from 2 columns to 13. The script was coded to take the 2nd column as the input for the moves.
Regards

Signed-off-by: mattis <mattis.backhaus@maxe-del.de>